Revamp Notepad Frame Layout and Introduce ButtonPlace Class
In this commit, the layout of the Frame widget housing the notes has undergone a significant transformation. The frame now spans the full width of the application, resulting in a more aesthetically pleasing and organized design.

To streamline the user interface, the "Mark Note" button has been removed, reducing unnecessary clutter and providing a cleaner overall appearance.

To enhance the flexibility of button placement, a new ButtonPlace class has been introduced. This class allows for the easy integration of buttons using the place method. Leveraging this class, the "Remove Note" button has been strategically placed in the bottom right corner of the notes. Furthermore, the button has been adorned with a GUI image, contributing to a visually appealing interface.

To improve the visibility of notes, the height of the Notes Frame has been increased, ensuring a more user-friendly experience.